\name{lerna}
\alias{lerna}
\docType{data}
\title{
  Stable isotopic data of the Lerna population
}
\description{
  This data set gives the stable carbons and nitrogen isotope ratios of the 39 human bone samples from Middle Bronze Age (ca. 2100--1700 BC) Lerna, Greece reported in Triantaphyllou et al. (2008).
}
\usage{lerna}
\format{
  \code{lerna} is a data frame with 39 individuals (rows) and 5 variables (columns) of ID (skeletal number), sex, estimated age at death, and carbon and nitrogen isotope ratios, named \code{id}, \code{sex}, \code{age}, \code{d13C}, and \code{d15N}, respectively. \code{sex} is a factor with levels \code{f}, \code{m}, and \code{u}, which corresponds to famale, male, unknown, respectively. The estimated ages at death were midpoint values. 
}
\details{
  See detail for Triantaphyllou et al. (2008).
}
\source{
  Triantaphyllou, S., Richards, M. P., Zerner, C., and Voutsaki, S. (2008). Isotopic dietary reconstruction of humans from Middle Bronze Age Lerna, Argolid, Greece. \emph{J. Archaeol. Sci.} \bold{35}, 3028--3034.
}
\keyword{datasets}

